Fixed issue with determining if value is a boolean
Issue with checking if the value is a boolean and setting that value into the json file. was skipping making sure that it added a true boolean value as "true" to the json doc as it was adding it as "True" and causing errors with  the json file when decoding it.